+++ tests/Feature/Request/RoaRequestTest.php
@@ -5,7 +5,6 @@ namespace AlibabaCloud\Client\Tests\Feature\Request;
 use AlibabaCloud\Client\AlibabaCloud;
 use AlibabaCloud\Client\Exception\ClientException;
 use AlibabaCloud\Client\Exception\ServerException;
-use AlibabaCloud\Client\Request\RoaRequest;
 use AlibabaCloud\Client\Tests\Mock\Services\Nlp\NlpRequest;
 use PHPUnit\Framework\TestCase;
 
@@ -74,41 +73,5 @@ class RoaRequestTest extends TestCase
                           ->timeout(30)
                           ->request();
         self::assertEquals('Iphone', $result['data'][0]['word']);
-    }
-
-    /**
-     * @throws ClientException
-     */
-    public function testCall()
-    {
-        $request = new RoaRequest();
-        self::assertEquals([], $request->pathParameters);
-
-        $request->setPrefix('set');
-        self::assertEquals('set', $request->getPrefix());
-        self::assertEquals(['Prefix' => 'set',], $request->pathParameters);
-
-        $request->withPrefix('with');
-        self::assertEquals('with', $request->getPrefix());
-        self::assertEquals(['Prefix' => 'with',], $request->pathParameters);
-
-        $request->setprefix('set');
-        self::assertEquals('set', $request->getprefix());
-        self::assertEquals(['Prefix' => 'with', 'prefix' => 'set',], $request->pathParameters);
-
-        $request->withprefix('with');
-        self::assertEquals('with', $request->getprefix());
-        self::assertEquals(['Prefix' => 'with', 'prefix' => 'with',], $request->pathParameters);
-    }
-
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessage Call to undefined method AlibabaCloud\Client\Request\RoaRequest::nowithvalue()
-     * @throws ClientException
-     */
-    public function testCallException()
-    {
-        $request = new RoaRequest();
-        $request->nowithvalue('value');
     }
 }
