+++ tests/Unit/Request/Traits/ClientTraitTest.php
@@ -1,8 +1,10 @@
 <?php
 
-namespace AlibabaCloud\Client\Tests\Unit\Request;
+namespace AlibabaCloud\Client\Tests\Unit\Request\Traits;
 
 use AlibabaCloud\Client\AlibabaCloud;
+use AlibabaCloud\Client\Clients\EcsRamRoleClient;
+use AlibabaCloud\Client\Credentials\Providers\CredentialsProvider;
 use AlibabaCloud\Client\Credentials\RamRoleArnCredential;
 use AlibabaCloud\Client\Credentials\Requests\AssumeRole;
 use AlibabaCloud\Client\Credentials\Requests\GenerateSessionAccessKey;
@@ -22,6 +24,66 @@ use PHPUnit\Framework\TestCase;
  */
 class ClientTraitTest extends TestCase
 {
+    /**
+     * @throws ClientException
+     * @throws ServerException
+     */
+    public function testCredential()
+    {
+        // Setup
+        $clientName = __METHOD__;
+
+        // Test
+        AlibabaCloud::accessKeyClient('key', 'secret')
+                    ->asDefaultClient()
+                    ->regionId('cn-hangzhou')
+                    ->name($clientName);
+
+        $request = (new DescribeCdnServiceRequest())->client($clientName);
+
+        // Assert
+        self::assertEquals('key', $request->credential()->getAccessKeyId());
+        self::assertEquals('secret', $request->credential()->getAccessKeySecret());
+    }
+
+    /**
+     * @throws ClientException
+     */
+    public function testHttpClientWithCustomChain()
+    {
+        $name = 'testHttpClientWithCustomChain';
+        AlibabaCloud::flush();
+        CredentialsProvider::chain(
+            function () use ($name) {
+                AlibabaCloud::ecsRamRoleClient('role')->name($name);
+            }
+        );
+        $request = AlibabaCloud::rpc()->client($name);
+        self::assertInstanceOf(EcsRamRoleClient::class, $request->httpClient());
+    }
+
+    /**
+     * @expectedException \AlibabaCloud\Client\Exception\ClientException
+     * @expectedExceptionMessage Client 'DefaultProvider' not found
+     * @throws ClientException
+     */
+    public function testHttpClientWithDefaultProvider()
+    {
+        CredentialsProvider::flush();
+        $request = AlibabaCloud::rpc()->client('DefaultProvider');
+        $request->httpClient();
+    }
+
+    /**
+     * @expectedException \AlibabaCloud\Client\Exception\ClientException
+     * @expectedExceptionMessage Client 'no' not found
+     * @throws ClientException
+     */
+    public function testHttpClient()
+    {
+        $request = AlibabaCloud::rpc()->client('no');
+        $request->httpClient();
+    }
 
     /**
      * @throws ClientException
@@ -59,28 +121,6 @@ class ClientTraitTest extends TestCase
 
         // Assert
         $this->assertEquals($expected, $request->options['headers']['client']);
-    }
-
-    /**
-     * @throws ClientException
-     * @throws ServerException
-     */
-    public function testCredential()
-    {
-        // Setup
-        $clientName = __METHOD__;
-
-        // Test
-        AlibabaCloud::accessKeyClient('key', 'secret')
-                    ->asDefaultClient()
-                    ->regionId('cn-hangzhou')
-                    ->name($clientName);
-
-        $request = (new DescribeCdnServiceRequest())->client($clientName);
-
-        // Assert
-        self::assertEquals('key', $request->credential()->getAccessKeyId());
-        self::assertEquals('secret', $request->credential()->getAccessKeySecret());
     }
 
     /**
